Refactor CustomDropsProvider to be more flexible
It is no longer an Integration, but requires a CustomDropsProviderFactory which is the Integration.

Providers can now return a Collection of ItemStacks instead of being limited to one. They also have access to the processing context.

This change breaks all CustomDropsProviders, built-in providers are updated but third party ones must update, they will throw exceptions otherwise.

The configuration of drops is also improved and is fortunately not broken. It allows providers to have their own configuration, much like anything else extensible in this plugin.